yaml2: Meta should keep defaults, and Res should have kind
This would previously panic since it wouldn't get a kind, and the meta
parameters would overwrite the defaults so it would block because limit
didn't have the default of +inf.

The removal of the SetKind was my fault in:

314c12f00c092f3a7af23d6cd1c75787d3e8fbaa

It's funny because it ends in `bad`. Guess I should have checked that!